dedupe tons of code w/ generic stream proxier
