+++ engine/buildkit/filesync.go
@@ -25,7 +25,6 @@ import (
 	bksolverpb "github.com/moby/buildkit/solver/pb"
 	"github.com/moby/buildkit/util/bklog"
 	filesynctypes "github.com/tonistiigi/fsutil/types"
-	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
@@ -350,8 +349,6 @@ func (c *Client) getLocalExportOpts(stream grpc.ServerStream) (context.Context, 
 	}, nil
 }
 
-// TODO: could make generic func to reduce tons of below boilerplate
-
 // for local dir imports
 type fileSyncServerProxy struct {
 	c *Client
@@ -372,62 +369,7 @@ func (p *fileSyncServerProxy) DiffCopy(stream filesync.FileSync_DiffCopyServer) 
 	if err != nil {
 		return err
 	}
-
-	var eg errgroup.Group
-	var done bool
-	eg.Go(func() (rerr error) {
-		defer func() {
-			diffCopyClient.CloseSend() // TODO: make sure all the other streams do this too, including non-filesync ones
-			if rerr == io.EOF {
-				rerr = nil
-			}
-			if errors.Is(rerr, context.Canceled) && done {
-				rerr = nil
-			}
-			if rerr != nil {
-				cancel()
-			}
-		}()
-		for {
-			pkt, err := withContext(ctx, func() (*filesynctypes.Packet, error) {
-				var pkt filesynctypes.Packet
-				err := stream.RecvMsg(&pkt)
-				return &pkt, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := diffCopyClient.SendMsg(pkt); err != nil {
-				return err
-			}
-		}
-	})
-	eg.Go(func() (rerr error) {
-		defer func() {
-			bklog.G(ctx).WithError(rerr).Debug("diffcopy sender done")
-			if rerr == io.EOF {
-				rerr = nil
-			}
-			if rerr == nil {
-				done = true
-			}
-			cancel()
-		}()
-		for {
-			pkt, err := withContext(ctx, func() (*filesynctypes.Packet, error) {
-				var pkt filesynctypes.Packet
-				err := diffCopyClient.RecvMsg(&pkt)
-				return &pkt, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := stream.SendMsg(pkt); err != nil {
-				return err
-			}
-		}
-	})
-	return eg.Wait()
+	return proxyStream[filesynctypes.Packet](ctx, diffCopyClient, stream)
 }
 
 func (p *fileSyncServerProxy) TarStream(stream filesync.FileSync_TarStreamServer) error {
@@ -441,56 +383,7 @@ func (p *fileSyncServerProxy) TarStream(stream filesync.FileSync_TarStreamServer
 	if err != nil {
 		return err
 	}
-
-	var eg errgroup.Group
-	eg.Go(func() (rerr error) {
-		defer func() {
-			tarStreamClient.CloseSend()
-			if rerr == io.EOF {
-				rerr = nil
-			}
-			if rerr != nil {
-				cancel()
-			}
-		}()
-		for {
-			pkt, err := withContext(ctx, func() (*filesynctypes.Packet, error) {
-				var pkt filesynctypes.Packet
-				err := stream.RecvMsg(&pkt)
-				return &pkt, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := tarStreamClient.SendMsg(pkt); err != nil {
-				return err
-			}
-		}
-	})
-	eg.Go(func() (rerr error) {
-		defer func() {
-			if rerr == io.EOF {
-				rerr = nil
-			}
-			if rerr != nil {
-				cancel()
-			}
-		}()
-		for {
-			pkt, err := withContext(ctx, func() (*filesynctypes.Packet, error) {
-				var pkt filesynctypes.Packet
-				err := tarStreamClient.RecvMsg(&pkt)
-				return &pkt, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := stream.SendMsg(pkt); err != nil {
-				return err
-			}
-		}
-	})
-	return eg.Wait()
+	return proxyStream[filesynctypes.Packet](ctx, tarStreamClient, stream)
 }
 
 // TODO: workaround needed until upstream fix: https://github.com/moby/buildkit/pull/4049
@@ -587,82 +480,9 @@ func (p *fileSendServerProxy) DiffCopy(stream filesync.FileSend_DiffCopyServer) 
 	}
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
-
-	var eg errgroup.Group
-	eg.Go(func() (rerr error) {
-		defer func() {
-			diffCopyClient.CloseSend()
-			if errors.Is(rerr, io.EOF) {
-				rerr = nil
-			}
-			if rerr != nil {
-				cancel()
-			}
-		}()
-		for {
-			msg, err := withContext(ctx, func() (any, error) {
-				var msg any = &filesynctypes.Packet{}
-				if useBytesMessageType {
-					msg = &filesync.BytesMessage{}
-				}
-				err := stream.RecvMsg(msg)
-				return msg, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := diffCopyClient.SendMsg(msg); err != nil {
-				return err
-			}
-		}
-	})
-	eg.Go(func() (rerr error) {
-		defer func() {
-			if errors.Is(rerr, io.EOF) {
-				rerr = nil
-			}
-			if rerr != nil {
-				cancel()
-			}
-		}()
-		for {
-			msg, err := withContext(ctx, func() (any, error) {
-				var msg any = &filesynctypes.Packet{}
-				if useBytesMessageType {
-					msg = &filesync.BytesMessage{}
-				}
-				err := diffCopyClient.RecvMsg(msg)
-				return msg, err
-			})
-			if err != nil {
-				return err
-			}
-			if err := stream.SendMsg(msg); err != nil {
-				return err
-			}
-		}
-	})
-	return eg.Wait()
-}
-
-// withContext adapts a blocking function to a context-aware function. It's
-// up to the caller to ensure that the blocking function f will unblock at
-// some time, otherwise there can be a goroutine leak.
-func withContext[T any](ctx context.Context, f func() (T, error)) (T, error) {
-	type result struct {
-		v   T
-		err error
-	}
-	ch := make(chan result, 1)
-	go func() {
-		v, err := f()
-		ch <- result{v, err}
-	}()
-	select {
-	case <-ctx.Done():
-		var zero T
-		return zero, ctx.Err()
-	case r := <-ch:
-		return r.v, r.err
+	if useBytesMessageType {
+		return proxyStream[filesync.BytesMessage](ctx, diffCopyClient, stream)
+	} else {
+		return proxyStream[filesynctypes.Packet](ctx, diffCopyClient, stream)
 	}
 }
